Move leaf hash methods to hash.go
